rcu: For RCU grace-period kthread starvation, dump last CPU it ran on
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Paul E. McKenney <paulmck@kernel.org>
commit 243027a3c80564bf96e40437ffac46efb9f5f2b5

When the RCU CPU stall-warning code detects that the RCU grace-period
kthread is being starved, it dumps that kthread's stack.  This can
sometimes be useful, but it is also useful to know what is running on the
CPU that this kthread is attempting to run on.  This commit therefore
adds a stack trace of this CPU in order to help track down whatever it
is that might be preventing RCU's grace-period kthread from running.

	Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
(cherry picked from commit 243027a3c80564bf96e40437ffac46efb9f5f2b5)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>