mm, memcg: decouple e{low,min} state mutations from protection checks
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Chris Down <chris@chrisdown.name>
commit 45c7f7e1ef17f09fe70bad4b705ce43772153fd7

mem_cgroup_protected currently is both used to set effective low and min
and return a mem_cgroup_protection based on the result.  As a user, this
can be a little unexpected: it appears to be a simple predicate function,
if not for the big warning in the comment above about the order in which
it must be executed.

This change makes it so that we separate the state mutations from the
actual protection checks, which makes it more obvious where we need to be
careful mutating internal state, and where we are simply checking and
don't need to worry about that.

[mhocko@suse.com - don't check protection on root memcgs]

	Suggested-by: Johannes Weiner <hannes@cmpxchg.org>
	Signed-off-by: Chris Down <chris@chrisdown.name>
	Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
	Acked-by: Johannes Weiner <hannes@cmpxchg.org>
	Acked-by: Michal Hocko <mhocko@suse.com>
	Cc: Roman Gushchin <guro@fb.com>
	Cc: Yafang Shao <laoar.shao@gmail.com>
Link: http://lkml.kernel.org/r/ff3f915097fcee9f6d7041c084ef92d16aaeb56a.1594638158.git.chris@chrisdown.name
	Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit 45c7f7e1ef17f09fe70bad4b705ce43772153fd7)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>