scsi: target: iscsi: Drop unnecessary container_of()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Guenter Roeck <linux@roeck-us.net>
commit 998da772fd86bd4ba2ba8c0ac2c9bba4815a952d

The structure pointer passed to container_of() is never NULL; that was
already checked. That means that the result of container_of() operations
on it is also never NULL, even though se_node_acl is the first element
of the structure embedding it. On top of that, it is misleading to perform
a NULL check on the result of container_of() because the position of the
contained element could change, which would make the test invalid.
Remove the unnecessary NULL check.

As it turns out, the container_of operation was only made for the purpose
of the NULL check. If the container_of is actually needed, it is repeated
later. Remove the container_of operation as well.

The NULL check was identified and removed with the following Coccinelle
script.

@@
type t;
identifier v;
statement s;
@@

<+...
(
  t v = container_of(...);
|
  v = container_of(...);
)
  ...
  when != v
- if (\( !v \| v == NULL \) ) s
...+>

Link: https://lore.kernel.org/r/20210510040817.2050266-1-linux@roeck-us.net
	Cc: "Martin K. Petersen" <martin.petersen@oracle.com>
	Cc: Hou Pu <houpu@bytedance.com>
	Cc: Mike Christie <michael.christie@oracle.com>
	Signed-off-by: Guenter Roeck <linux@roeck-us.net>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 998da772fd86bd4ba2ba8c0ac2c9bba4815a952d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>