softirq: Don't skip softirq execution when softirq thread is parking
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Matthias Kaehlcke <mka@chromium.org>
commit 1342d8080f6183b0419a9246c6e6545e21fa1e05

When a CPU is unplugged the kernel threads of this CPU are parked (see
smpboot_park_threads()). kthread_park() is used to mark each thread as
parked and wake it up, so it can complete the process of parking itselfs
(see smpboot_thread_fn()).

If local softirqs are pending on interrupt exit invoke_softirq() is called
to process the softirqs, however it skips processing when the softirq
kernel thread of the local CPU is scheduled to run. The softirq kthread is
one of the threads that is parked when a CPU is unplugged. Parking the
kthread wakes it up, however only to complete the parking process, not to
process the pending softirqs. Hence processing of softirqs at the end of an
interrupt is skipped, but not done elsewhere, which can result in warnings
about pending softirqs when a CPU is unplugged:

/sys/devices/system/cpu # echo 0 > cpu4/online
[ ... ] NOHZ: local_softirq_pending 02
[ ... ] NOHZ: local_softirq_pending 202
[ ... ] CPU4: shutdown
[ ... ] psci: CPU4 killed.

Don't skip processing of softirqs at the end of an interrupt when the
softirq thread of the CPU is parking.

	Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
	Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
	Cc: Peter Zijlstra <peterz@infradead.org>
	Cc: Steven Rostedt <rostedt@goodmis.org>
	Cc: "Paul E . McKenney" <paulmck@linux.vnet.ibm.com>
	Cc: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
	Cc: Douglas Anderson <dianders@chromium.org>
	Cc: Stephen Boyd <swboyd@chromium.org>
Link: https://lkml.kernel.org/r/20190128234625.78241-3-mka@chromium.org

(cherry picked from commit 1342d8080f6183b0419a9246c6e6545e21fa1e05)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>