locking/rtmutex: Implement equal priority lock stealing
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Gregory Haskins <ghaskins@novell.com>
commit 48eb3f4fcfd35495a8357459aa6fe437aa430b00

The current logic only allows lock stealing to occur if the current task is
of higher priority than the pending owner.

Significant throughput improvements can be gained by allowing the lock
stealing to include tasks of equal priority when the contended lock is a
spin_lock or a rw_lock and the tasks are not in a RT scheduling task.

The assumption was that the system will make faster progress by allowing
the task already on the CPU to take the lock rather than waiting for the
system to wake up a different task.

This does add a degree of unfairness, but in reality no negative side
effects have been observed in the many years that this has been used in the
RT kernel.

[ tglx: Refactored and rewritten several times by Steve Rostedt, Sebastian
  	Siewior and myself ]

	Signed-off-by: Gregory Haskins <ghaskins@novell.com>
	Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
	Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
	Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lore.kernel.org/r/20210815211305.857240222@linutronix.de
(cherry picked from commit 48eb3f4fcfd35495a8357459aa6fe437aa430b00)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>