tracing: Quiet smp_processor_id() use in preemptable warning in hwlat
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Steven Rostedt (VMware) <rostedt@goodmis.org>
commit 51397dc6f283bb570e1cf8226017d300d8ea1f5b

The hardware latency detector (hwlat) has a mode that it runs one thread
across CPUs. The logic to move from the currently running CPU to the next
one in the list does a smp_processor_id() to find where it currently is.
Unfortunately, it's done with preemption enabled, and this triggers a
warning for using smp_processor_id() in a preempt enabled section.

As it is only using smp_processor_id() to get information on where it
currently is in order to simply move it to the next CPU, it doesn't really
care if it got moved in the mean time. It will simply balance out later if
such a case arises.

Switch smp_processor_id() to raw_smp_processor_id() to quiet that warning.

Link: https://lkml.kernel.org/r/20210804141848.79edadc0@oasis.local.home

	Acked-by: Daniel Bristot de Oliveira <bristot@redhat.com>
Fixes: 8fa826b7344d ("trace/hwlat: Implement the mode config option")
	Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 51397dc6f283bb570e1cf8226017d300d8ea1f5b)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>