perf build: Regenerate the FEATURE_DUMP file after extra feature checks
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Jiri Olsa <jolsa@redhat.com>
commit fbed59f844912f377b83cc25594c692b5f6ebae2

Feature detection is done in tools/build/Makefile.feature, we may exit
there with some features not detected and then, in
tools/perf/Makefile.config try adding extra libraries to link and then
do extra feature checks to see if we now find the feature.

This is the case with the disassembler-four-args that checks if the
diassembler() function in libopcodes (binutils) has a signature with
one or with four arguments, as this is not ABI and they changed it at
some point.

This is not a problem when doing normal builds, for instance:

  $ make -C tools/perf O=/tmp/build/perf

As we don't use what is in FEATURE-DUMP at that point, but is a problem
if we pass FEATURE_DUMP=/previously-detected-features as we do in
'make -C tools/perf build-test' to reuse the feature detection in the
many build combinations we test there.

When that is done feature-disassembler-four-args will be set to 0, but
opensuse 15.1 has the four arguments function signature in
disassembler(). The build thus fails.

Fix it by rewriting the FEATURE-DUMP file at the end of
tools/perf/Makefile.config to register features we retested in that make
file.

	Signed-off-by: Jiri Olsa <jolsa@redhat.com>
	Reported-by: Arnaldo Carvalho de Melo <acme@redhat.com>
	Tested-by: Arnaldo Carvalho de Melo <acme@redhat.com>
	Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>
(cherry picked from commit fbed59f844912f377b83cc25594c692b5f6ebae2)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>