KVM: s390: Enable specification exception interpretation
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Janis Schoetterl-Glausch <scgl@linux.ibm.com>
commit 7119decf47d9867266459615be502e5d2cecedba

When this feature is enabled the hardware is free to interpret
specification exceptions generated by the guest, instead of causing
program interruption interceptions.

This benefits (test) programs that generate a lot of specification
exceptions (roughly 4x increase in exceptions/sec).

Interceptions will occur as before if ICTL_PINT is set,
i.e. if guest debug is enabled.

There is no indication if this feature is available or not and the
hardware is free to interpret or not. So we can simply set this bit and
if the hardware ignores it we fall back to intercept 8 handling.

	Signed-off-by: Janis Schoetterl-Glausch <scgl@linux.ibm.com>
Link: https://lore.kernel.org/linux-s390/20210706114714.3936825-1-scgl@linux.ibm.com/
	Reviewed-by: David Hildenbrand <david@redhat.com>
	Reviewed-by: Christian Borntraeger <borntraeger@de.ibm.com>
	Signed-off-by: Christian Borntraeger <borntraeger@de.ibm.com>
(cherry picked from commit 7119decf47d9867266459615be502e5d2cecedba)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>