pinctrl: intel: check REVID register value for device presence
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Roger Pau Monne <roger.pau@citrix.com>
commit 39c1f1bd8adf29840ef52225ce95f390e3d7b2cf

Use the value read from the REVID register in order to check for the
presence of the device. A read of all ones is treated as if the device
is not present, and hence probing is ended.

This fixes an issue when running as a Xen PVH dom0, where the ACPI
DSDT table is provided unmodified to dom0 and hence contains the
pinctrl devices, but the MMIO region(s) containing the device
registers might not be mapped in the guest physical memory map if such
region(s) are not exposed on a PCI device BAR or marked as reserved in
the host memory map.

Fixes: 91d898e51e60 ("pinctrl: intel: Convert capability list to features")
	Suggested-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
	Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
	Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
(cherry picked from commit 39c1f1bd8adf29840ef52225ce95f390e3d7b2cf)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>