scsi: scsi_dh_alua: Retry RTPG on a different path after failure
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Martin Wilck <mwilck@suse.com>
commit ee8868c5c78f16fb726775741aeab8a233373332

If an RTPG fails, we can't infer anything wrt. the state of the ports in
the port group except that we were unable to reach the one port on which
the RTPG had failed. "offline" is just a secondary port state, which means
that we can't infer the state of any port in the PG from the failure (in
fact, even the failed port might still be in "active/optimized" primary
port access state).

Therefore, when we encounter an RTPG failure, we should retry the RTPG on a
different port. This avoids falsely setting port states to offline for
unreachable ports. To do this, ports on which an RTPG has failed are
temporarily set to "disabled" to avoid repeating the failed I/O on the same
target port. Once the RTPG has either succeeded on one port or failed on
all ports of the PG, the ports are enabled again.

Link: https://lore.kernel.org/r/20210514153214.5626-1-mwilck@suse.com
	Signed-off-by: Martin Wilck <mwilck@suse.com>
	Signed-off-by: Hannes Reinecke <hare@suse.de>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit ee8868c5c78f16fb726775741aeab8a233373332)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>