scsi: iscsi: Delay freeing target_id
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Mike Christie <michael.christie@oracle.com>
commit e463f96bdc975b780161541877f521da078a8143

If we are doing async removal of the session, we could be doing a
scsi_remove_target from the removal workqueue, and for the offload case we
could be doing a new session addition and scan to the same host. The
add/scan might then end up trying to use the target_id of the target we are
removing.

This patch just has a delay the freeing of the target_id until after the
scsi_remove_target has completed, so we know it's no longer in use.

Link: https://lore.kernel.org/r/1593632868-6808-2-git-send-email-michael.christie@oracle.com
	Reviewed-by: Lee Duncan <lduncan@suse.com>
	Signed-off-by: Mike Christie <michael.christie@oracle.com>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit e463f96bdc975b780161541877f521da078a8143)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>