scsi: iscsi: Have abort handler get ref to conn
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Mike Christie <michael.christie@oracle.com>
commit d39df158518ccc3bf24ee18082b5e100c8f014aa

If SCSI midlayer is aborting a task when we are tearing down the conn we
could free the conn while the abort thread is accessing the conn. This has
the abort handler get a ref to the conn so it won't be freed from under it.

Note: this is not needed for device/target reset because we are holding the
eh_mutex when accessing the conn.

Link: https://lore.kernel.org/r/20210525181821.7617-12-michael.christie@oracle.com
	Reviewed-by: Lee Duncan <lduncan@suse.com>
	Signed-off-by: Mike Christie <michael.christie@oracle.com>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit d39df158518ccc3bf24ee18082b5e100c8f014aa)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>