scsi: iscsi: Fix shost->max_id use
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Mike Christie <michael.christie@oracle.com>
commit bdd4aad7ff92ae39c2e93c415bb6761cb8b584da

The iscsi offload drivers are setting the shost->max_id to the max number
of sessions they support. The problem is that max_id is not the max number
of targets but the highest identifier the targets can have. To use it to
limit the number of targets we need to set it to max sessions - 1, or we
can end up with a session we might not have preallocated resources for.

Link: https://lore.kernel.org/r/20210525181821.7617-15-michael.christie@oracle.com
	Reviewed-by: Lee Duncan <lduncan@suse.com>
	Signed-off-by: Mike Christie <michael.christie@oracle.com>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit bdd4aad7ff92ae39c2e93c415bb6761cb8b584da)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>