x86/fpu: Rename copy_fpregs_to_fpstate() to save_fpregs_to_fpstate()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Thomas Gleixner <tglx@linutronix.de>
commit ebe7234b08a42d69bae94c4062a84777ea26ef99

A copy is guaranteed to leave the source intact, which is not the case when
FNSAVE is used as that reinitilizes the registers.

Save does not make such guarantees and it matches what this is about,
i.e. to save the state for a later restore.

Rename it to save_fpregs_to_fpstate().

	Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
	Signed-off-by: Borislav Petkov <bp@suse.de>
	Reviewed-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210623121454.508853062@linutronix.de
(cherry picked from commit ebe7234b08a42d69bae94c4062a84777ea26ef99)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>