block, bfq: boost throughput by extending queue-merging times
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Pietro Pedroni <pedroni.pietro.96@gmail.com>
commit 7812472f973047a886e4ed9a91d98d6627dd746f

One of the methods with which bfq boosts throughput is by merging queues.
One of the merging variants in bfq is the stable merge.
This mechanism is activated between two queues only if they are created
within a certain maximum time T1 from each other.
Merging can happen soon or be delayed. In the second case, before
merging, bfq needs to evaluate a throughput-boost parameter that
indicates whether the queue generates a high throughput is served alone.
Merging occurs when this throughput-boost is not high enough.
In particular, this parameter is evaluated and late merging may occur
only after at least a time T2 from the creation of the queue.

Currently T1 and T2 are set to 180ms and 200ms, respectively.
In this way the merging mechanism rarely occurs because time is not
enough. This results in a noticeable lowering of the overall throughput
with some workloads (see the example below).

This commit introduces two constants bfq_activation_stable_merging and
bfq_late_stable_merging in order to increase the duration of T1 and T2.
Both the stable merging activation time and the late merging
time are set to 600ms. This value has been experimentally evaluated
using sqlite benchmark in the Phoronix Test Suite on a HDD.
The duration of the benchmark before this fix was 111.02s, while now
it has reached 97.02s, a better result than that of all the other
schedulers.

	Signed-off-by: Pietro Pedroni <pedroni.pietro.96@gmail.com>
	Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Link: https://lore.kernel.org/r/20210619140948.98712-5-paolo.valente@linaro.org
	Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit 7812472f973047a886e4ed9a91d98d6627dd746f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>