Bluetooth: btbcm: Obtain and handle reset GPIO
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Linus Walleij <linus.walleij@linaro.org>
commit 2f40796671f03c0eb193f44c7d4a6855ab9b30f6

Obtain and drive the optional reset GPIO line if this is
not hardwired in the platform. This is needed on the
Samsung GT-I9070 mobile phone.

The invers of power is used, this should be OK to apply
simultaneously as the power signal according to
figures 47-51 on pages 159-161 in the BCM4330 datasheet.

	Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
	Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 2f40796671f03c0eb193f44c7d4a6855ab9b30f6)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>