iommu/vt-d: Remove redundant assignment to variable agaw
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Colin Ian King <colin.king@canonical.com>
commit 05d2cbf969be465027d645f855419dbe49ded849

The variable agaw is initialized with a value that is never read and it
is being updated later with a new value as a counter in a for-loop. The
initialization is redundant and can be removed.

Addresses-Coverity: ("Unused value")
	Signed-off-by: Colin Ian King <colin.king@canonical.com>
	Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Link: https://lore.kernel.org/r/20210416171826.64091-1-colin.king@canonical.com
Link: https://lore.kernel.org/r/20210610020115.1637656-2-baolu.lu@linux.intel.com
	Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit 05d2cbf969be465027d645f855419dbe49ded849)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>