iommu/amd: Tailored gather logic for AMD
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Nadav Amit <namit@vmware.com>
commit fe6d269d0e9b05fa5233f810a3cb7cce4077b261

AMD's IOMMU can flush efficiently (i.e., in a single flush) any range.
This is in contrast, for instnace, to Intel IOMMUs that have a limit on
the number of pages that can be flushed in a single flush.  In addition,
AMD's IOMMU do not care about the page-size, so changes of the page size
do not need to trigger a TLB flush.

So in most cases, a TLB flush due to disjoint range is not needed for
AMD. Yet, vIOMMUs require the hypervisor to synchronize the virtualized
IOMMU's PTEs with the physical ones. This process induce overheads, so
it is better not to cause unnecessary flushes, i.e., flushes of PTEs
that were not modified.

Implement and use amd_iommu_iotlb_gather_add_page() and use it instead
of the generic iommu_iotlb_gather_add_page(). Ignore disjoint regions
unless "non-present cache" feature is reported by the IOMMU
capabilities, as this is an indication we are running on a physical
IOMMU. A similar indication is used by VT-d (see "caching mode"). The
new logic retains the same flushing behavior that we had before the
introduction of page-selective IOTLB flushes for AMD.

On virtualized environments, check if the newly flushed region and the
gathered one are disjoint and flush if it is.

	Cc: Joerg Roedel <joro@8bytes.org>
	Cc: Will Deacon <will@kernel.org>
	Cc: Jiajun Cao <caojiajun@vmware.com>
	Cc: Lu Baolu <baolu.lu@linux.intel.com>
	Cc: iommu@lists.linux-foundation.org
	Cc: linux-kernel@vger.kernel.org>
	Reviewed-by: Robin Murphy <robin.murphy@arm.com>
	Signed-off-by: Nadav Amit <namit@vmware.com>
Link: https://lore.kernel.org/r/20210723093209.714328-6-namit@vmware.com
	Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit fe6d269d0e9b05fa5233f810a3cb7cce4077b261)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>