xfs: scrub should mark a directory corrupt if any entries cannot be iget'd
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Darrick J. Wong <darrick.wong@oracle.com>
commit 4b80ac64450f169bae364df631d233d66070a06e

It's possible that xfs_iget can return EINVAL for inodes that the inobt
thinks are free, or ENOENT for inodes that look free.  If this is the
case, mark the directory corrupt immediately when we check ftype.  Note
that we already check the ftype of the '.' and '..' entries, so we
can skip the iget part since we already know the inode type for '.' and
we have a separate parent pointer scrubber for '..'.

Fixes: a5c46e5e8912 ("xfs: scrub directory metadata")
	Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
	Reviewed-by: Christoph Hellwig <hch@lst.de>
(cherry picked from commit 4b80ac64450f169bae364df631d233d66070a06e)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>