ima: Don't modify file descriptor mode on the fly
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Roberto Sassu <roberto.sassu@huawei.com>
commit 207cdd565dfc95a0a5185263a567817b7ebf5467

Commit a408e4a86b36b ("ima: open a new file instance if no read
permissions") already introduced a second open to measure a file when the
original file descriptor does not allow it. However, it didn't remove the
existing method of changing the mode of the original file descriptor, which
is still necessary if the current process does not have enough privileges
to open a new one.

Changing the mode isn't really an option, as the filesystem might need to
do preliminary steps to make the read possible. Thus, this patch removes
the code and keeps the second open as the only option to measure a file
when it is unreadable with the original file descriptor.

	Cc: <stable@vger.kernel.org> # 4.20.x: 0014cc04e8ec0 ima: Set file->f_mode
Fixes: 2fe5d6def1672 ("ima: integrity appraisal extension")
	Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>
	Reviewed-by: Christoph Hellwig <hch@lst.de>
	Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>
(cherry picked from commit 207cdd565dfc95a0a5185263a567817b7ebf5467)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>