KVM: s390: clear kicked_mask before sleeping again
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Halil Pasic <pasic@linux.ibm.com>
commit 9b57e9d5010bbed7c0d9d445085840f7025e6f9a

The idea behind kicked mask is that we should not re-kick a vcpu that
is already in the "kick" process, i.e. that was kicked and is
is about to be dispatched if certain conditions are met.

The problem with the current implementation is, that it assumes the
kicked vcpu is going to enter SIE shortly. But under certain
circumstances, the vcpu we just kicked will be deemed non-runnable and
will remain in wait state. This can happen, if the interrupt(s) this
vcpu got kicked to deal with got already cleared (because the interrupts
got delivered to another vcpu). In this case kvm_arch_vcpu_runnable()
would return false, and the vcpu would remain in kvm_vcpu_block(),
but this time with its kicked_mask bit set. So next time around we
wouldn't kick the vcpu form __airqs_kick_single_vcpu(), but would assume
that we just kicked it.

Let us make sure the kicked_mask is cleared before we give up on
re-dispatching the vcpu.

Fixes: 9f30f6216378 ("KVM: s390: add gib_alert_irq_handler()")
	Reported-by: Matthew Rosato <mjrosato@linux.ibm.com>
	Signed-off-by: Halil Pasic <pasic@linux.ibm.com>
	Reviewed-by: Christian Borntraeger <borntraeger@de.ibm.com>
	Reviewed-by: Michael Mueller <mimu@linux.ibm.com>
	Reviewed-by: Claudio Imbrenda <imbrenda@linux.ibm.com>
Link: https://lore.kernel.org/r/20211019175401.3757927-2-pasic@linux.ibm.com
	Signed-off-by: Christian Borntraeger <borntraeger@de.ibm.com>
(cherry picked from commit 9b57e9d5010bbed7c0d9d445085840f7025e6f9a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>