KVM: s390: pv: avoid stalls when making pages secure
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Claudio Imbrenda <imbrenda@linux.ibm.com>
commit f0a1a0615a6ff6d38af2c65a522698fb4bb85df6

Improve make_secure_pte to avoid stalls when the system is heavily
overcommitted. This was especially problematic in kvm_s390_pv_unpack,
because of the loop over all pages that needed unpacking.

Due to the locks being held, it was not possible to simply replace
uv_call with uv_call_sched. A more complex approach was
needed, in which uv_call is replaced with __uv_call, which does not
loop. When the UVC needs to be executed again, -EAGAIN is returned, and
the caller (or its caller) will try again.

When -EAGAIN is returned, the path is the same as when the page is in
writeback (and the writeback check is also performed, which is
harmless).

Fixes: 214d9bbcd3a672 ("s390/mm: provide memory management functions for protected KVM guests")
	Signed-off-by: Claudio Imbrenda <imbrenda@linux.ibm.com>
	Reviewed-by: Janosch Frank <frankja@linux.ibm.com>
	Reviewed-by: Christian Borntraeger <borntraeger@de.ibm.com>
Link: https://lore.kernel.org/r/20210920132502.36111-5-imbrenda@linux.ibm.com
	Signed-off-by: Christian Borntraeger <borntraeger@de.ibm.com>
(cherry picked from commit f0a1a0615a6ff6d38af2c65a522698fb4bb85df6)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>