Driver core: platform: Add extra error check in devm_platform_get_irqs_affinity()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author John Garry <john.garry@huawei.com>
commit 29f7c54b253fc18bff9bf7e9f303b75deb285c7a

The current check of nvec < minvec for nvec returned from
platform_irq_count() will not detect a negative error code in nvec.

This is because minvec is unsigned, and, as such, nvec is promoted to
unsigned in that check, which will make it a huge number (if it contained
-EPROBE_DEFER).

In practice, an error should not occur in nvec for the only in-tree
user, but add a check anyway.

Fixes: e15f2fa959f2 ("driver core: platform: Add devm_platform_get_irqs_affinity()")
	Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
	Signed-off-by: John Garry <john.garry@huawei.com>
Link: https://lore.kernel.org/r/1608561055-231244-1-git-send-email-john.garry@huawei.com
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 29f7c54b253fc18bff9bf7e9f303b75deb285c7a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>