RDMA/rxe: Protext kernel index from user space
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Bob Pearson <rpearsonhpe@gmail.com>
commit 5bcf5a59c41e19141783c7305d420a5e36c937b2

In order to prevent user space from modifying the index that belongs to
the kernel for shared queues let the kernel use a local copy of the index
and copy any new values of that index to the shared rxe_queue_bus struct.

This adds more switch statements which decreases the performance of the
queue API. Move the type into the parameter list for these functions so
that the compiler can optimize out the switch statements when the explicit
type is known. Modify all the calls in the driver on performance paths to
pass in the explicit queue type.

Link: https://lore.kernel.org/r/20210527194748.662636-4-rpearsonhpe@gmail.com
Link: https://lore.kernel.org/linux-rdma/20210526165239.GP1002214@@nvidia.com/
	Signed-off-by: Bob Pearson <rpearsonhpe@gmail.com>
	Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
(cherry picked from commit 5bcf5a59c41e19141783c7305d420a5e36c937b2)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>