perf arm-spe: Correct sample flags for SPE event
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Leo Yan <leo.yan@linaro.org>
commit 2f021954952f105e32383103cac06a095d9a1db6

Now it's hard code to set sample flags for CPU, TIME and TID for SPE
event, which is pointless.

The CPU is useful for sampling only for per-mmap case, it is used to
indicate the AUX trace is associated to which CPU.

The TIME sample is not needed for AUX event, since the time for AUX
event is not really used and this time is a different thing from the
timestamp in Arm SPE trace, the timestamp tracing which is controlled
by Arm SPE's config bit.

The TID sample is not useful for AUX event.

This patch corrects the sample flags for SPE event, it only set CPU
sample bit for per-cpu mmap case.

	Signed-off-by: Leo Yan <leo.yan@linaro.org>
	Reviewed-by: James Clark <james.clark@arm.com>
	Tested-by: James Clark <james.clark@arm.com>
	Cc: Alexander Shishkin <alexander.shishkin@linux.intel.com>
	Cc: Al Grant <Al.Grant@arm.com>
	Cc: Ingo Molnar <mingo@redhat.com>
	Cc: Jiri Olsa <jolsa@redhat.com>
	Cc: John Garry <john.garry@huawei.com>
	Cc: Mark Rutland <mark.rutland@arm.com>
	Cc: Mathieu Poirier <mathieu.poirier@linaro.org>
	Cc: Namhyung Kim <namhyung@kernel.org>
	Cc: Peter Zijlstra <peterz@infradead.org>
	Cc: Will Deacon <will@kernel.org>
	Cc: linux-arm-kernel@lists.infradead.org
Link: https://lore.kernel.org/r/20210519041546.1574961-2-leo.yan@linaro.org
	Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>
(cherry picked from commit 2f021954952f105e32383103cac06a095d9a1db6)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>