perf probe: Support probes on init functions for offline kernel
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Masami Hiramatsu <mhiramat@kernel.org>
commit f338de22192085a476dd1bc9eac5e074a48f34f9

'perf probe' internally checks the probe target is in the text area in
post-process (after analyzing debuginfo). But it fails if the probe
target is in the "inittext".

This is a good limitation for the online kernel because such functions
have gone after booting. However, for using it for boot-time tracing,
user may want to put a probe on init functions.

This skips the post checking process if the target is offline kenrel so
that user can get the probe definition on the init functions.

Without this patch:

  $ perf probe -k ./build-x86_64/vmlinux -D do_mount_root:10
  Probe point 'do_mount_root:10' not found.
    Error: Failed to add events.

With this patch:

  $ perf probe -k ./build-x86_64/vmlinux -D do_mount_root:10
  p:probe/do_mount_root_L10 mount_block_root+300

	Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>
	Cc: Alexander Shishkin <alexander.shishkin@linux.intel.com>
	Cc: Jiri Olsa <jolsa@redhat.com>
	Cc: Mark Rutland <mark.rutland@arm.com>
	Cc: Namhyung Kim <namhyung@kernel.org>
	Cc: Peter Zijlstra <peterz@infradead.org>
Link: http://lore.kernel.org/lkml/162282410293.452340.13347006295826431632.stgit@devnote2
	Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>
(cherry picked from commit f338de22192085a476dd1bc9eac5e074a48f34f9)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>