ibmvnic: don't stop queue in xmit
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Sukadev Bhattiprolu <sukadev@linux.ibm.com>
commit 8878e46fcfd46b19964bd90e13b25dd94cbfc9be

If adapter's resetting bit is on, discard the packet but don't stop the
transmit queue - instead leave that to the reset code. With this change,
it is possible that we may get several calls to ibmvnic_xmit() that simply
discard packets and return.

But if we stop the queue here, we might end up doing so just after
__ibmvnic_open() started the queues (during a hard/soft reset) and before
the ->resetting bit was cleared. If that happens, there will be no one to
restart queue and transmissions will be blocked indefinitely.

This can cause a TIMEOUT reset and with auto priority failover enabled,
an unnecessary FAILOVER reset to less favored backing device and then a
FAILOVER back to the most favored backing device. If we hit the window
repeatedly, we can get stuck in a loop of TIMEOUT, FAILOVER, FAILOVER
resets leaving the adapter unusable for extended periods of time.

Fixes: 7f5b030830fe ("ibmvnic: Free skb's in cases of failure in transmit")
	Reported-by: Abdul Haleem <abdhalee@in.ibm.com>
	Reported-by: Vaishnavi Bhat <vaish123@in.ibm.com>
	Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.ibm.com>
	Reviewed-by: Dany Madden <drt@linux.ibm.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 8878e46fcfd46b19964bd90e13b25dd94cbfc9be)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>