ipvs: fix the connection sync failed in some cases
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author guodeqing <geffrey.guo@huawei.com>
commit 8210e344ccb798c672ab237b1a4f241bda08909b

The sync_thread_backup only checks sk_receive_queue is empty or not,
there is a situation which cannot sync the connection entries when
sk_receive_queue is empty and sk_rmem_alloc is larger than sk_rcvbuf,
the sync packets are dropped in __udp_enqueue_schedule_skb, this is
because the packets in reader_queue is not read, so the rmem is
not reclaimed.

Here I add the check of whether the reader_queue of the udp sock is
empty or not to solve this problem.

Fixes: 2276f58ac589 ("udp: use a separate rx queue for packet reception")
	Reported-by: zhouxudong <zhouxudong8@huawei.com>
	Signed-off-by: guodeqing <geffrey.guo@huawei.com>
	Acked-by: Julian Anastasov <ja@ssi.bg>
	Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
(cherry picked from commit 8210e344ccb798c672ab237b1a4f241bda08909b)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>