selftests/bpf: Propagate error code of the command to vmtest.sh
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author KP Singh <kpsingh@google.com>
commit 2854436612c4d2606c9246ce2976ab6634276337

When vmtest.sh ran a command in a VM, it did not record or propagate the
error code of the command. This made the script less "script-able". The
script now saves the error code of the said command in a file in the VM,
copies the file back to the host and (when available) uses this error
code instead of its own.

	Signed-off-by: KP Singh <kpsingh@google.com>
	Signed-off-by: Andrii Nakryiko <andrii@kernel.org>
Link: https://lore.kernel.org/bpf/20210225161947.1778590-1-kpsingh@kernel.org
(cherry picked from commit 2854436612c4d2606c9246ce2976ab6634276337)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>