selftests/bpf: Rewrite test_tc_redirect.sh as prog_tests/tc_redirect.c
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Jussi Maki <joamaki@gmail.com>
commit 096eccdef0b32f47e9354231ddc3aaaf9527d51c

As discussed in [0], this ports test_tc_redirect.sh to the test_progs
framework and removes the old test.

This makes it more in line with rest of the tests and makes it possible
to run this test case with vmtest.sh and under the bpf CI.

The upcoming skb_change_head() helper fix in [0] is depending on it and
extending the test case to redirect a packet from L3 device to veth.

  [0] https://lore.kernel.org/bpf/20210427135550.807355-1-joamaki@gmail.com

	Signed-off-by: Jussi Maki <joamaki@gmail.com>
	Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Link: https://lore.kernel.org/bpf/20210505085925.783985-1-joamaki@gmail.com
(cherry picked from commit 096eccdef0b32f47e9354231ddc3aaaf9527d51c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>