scsi: ibmvfc: Fix up duplicate response detection
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Brian King <brking@linux.vnet.ibm.com>
commit e20f80b9b163dc402dca115eed0affba6df5ebb5

Commit a264cf5e81c7 ("scsi: ibmvfc: Fix command state accounting and stale
response detection") introduced a regression in detecting duplicate
responses. This was observed in test where a command was sent to the VIOS
and completed before ibmvfc_send_event() set the active flag to 1, which
resulted in the atomic_dec_if_positive() call in ibmvfc_handle_crq()
thinking this was a duplicate response, which resulted in scsi_done() not
getting called, so we then hit a SCSI command timeout for this command once
the timeout expires.  This simply ensures the active flag gets set prior to
making the hcall to send the command to the VIOS, in order to close this
window.

Link: https://lore.kernel.org/r/20211019152129.16558-1-brking@linux.vnet.ibm.com
Fixes: a264cf5e81c7 ("scsi: ibmvfc: Fix command state accounting and stale response detection")
	Cc: stable@vger.kernel.org
	Acked-by: Tyrel Datwyler <tyreld@linux.ibm.com>
	Signed-off-by: Brian King <brking@linux.vnet.ibm.com>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit e20f80b9b163dc402dca115eed0affba6df5ebb5)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>