ipv4: Allow sending multicast packets on specific i/f using VRF socket
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Robert Shearman <rshearma@vyatta.att-mail.com>
commit 854da991733d1b4f60042423875e32be7f8f0421

It is useful to be able to use the same socket for listening in a
specific VRF, as for sending multicast packets out of a specific
interface. However, the bound device on the socket currently takes
precedence and results in the packets not being sent.

Relax the condition on overriding the output interface to use for
sending packets out of UDP, raw and ping sockets to allow multicast
packets to be sent using the specified multicast interface.

	Signed-off-by: Robert Shearman <rshearma@vyatta.att-mail.com>
	Signed-off-by: Mike Manning <mmanning@vyatta.att-mail.com>
	Reviewed-by: David Ahern <dsahern@gmail.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 854da991733d1b4f60042423875e32be7f8f0421)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>