s390: mm: Fix secure storage access exception handling
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Janosch Frank <frankja@linux.ibm.com>
commit 85b18d7b5e7ffefb2f076186511d39c4990aa005

Turns out that the bit 61 in the TEID is not always 1 and if that's
the case the address space ID and the address are
unpredictable. Without an address and its address space ID we can't
export memory and hence we can only send a SIGSEGV to the process or
panic the kernel depending on who caused the exception.

Unfortunately bit 61 is only reliable if we have the "misc" UV feature
bit.

	Signed-off-by: Janosch Frank <frankja@linux.ibm.com>
	Reviewed-by: Christian Borntraeger <borntraeger@de.ibm.com>
Fixes: 084ea4d611a3d ("s390/mm: add (non)secure page access exceptions handlers")
	Cc: stable@vger.kernel.org
	Signed-off-by: Vasily Gorbik <gor@linux.ibm.com>
(cherry picked from commit 85b18d7b5e7ffefb2f076186511d39c4990aa005)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>