mm: prevent a warning when casting void* -> enum
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-358.el8
commit-author Palmer Dabbelt <palmerdabbelt@google.com>
commit 4708f31885a0d3e41a03c39501df697e73406f6a

I recently build the RISC-V port with LLVM trunk, which has introduced a
new warning when casting from a pointer to an enum of a smaller size.
This patch simply casts to a long in the middle to stop the warning.  I'd
be surprised this is the only one in the kernel, but it's the only one I
saw.

	Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>
	Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
	Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Link: http://lkml.kernel.org/r/20200227211741.83165-1-palmer@dabbelt.com
	Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit 4708f31885a0d3e41a03c39501df697e73406f6a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>